secondnode to easily attach to an external dockerhosts 'firstnode'
Use this container when you want to addnode to an external firstnode. Since wallet is linked to 'firstnode' you cannot start it without a firstnode running. Secondnode is intended to be a mix of firstnode & wallet features. secondnode is persistent with docker volumes, so you don't immediately lose your wallet.dat or blockchain sync by restarting the container.  However it does not contain its own miner like firstnode does, so it will not mine the blockchain. The secondnode however does not use container linking to attach to firstnode or wallet, so it can be started without any other container running already. Secondnode will also expose ports 20114 and 20115 to the dockerhost, so could be mapped to and reached from outside the docker network.